# openspec/specs/provider-switching/contracts/api.openapi.yaml
...
openapi: 3.1.0
info:
  title: Coffer — Provider Switching API
  version: 0.1.0
  description: The Provider Switching routes of Coffer's management API, generated from the daemon's Pydantic models. The behaviour behind them is this capability's spec.md; authentication, the error envelope and tracing are described once in the docs-site REST API reference. Every route except `GET /api/v1/daemon/status` requires the `X-Coffer-Token` header.
servers:
- url: http://127.0.0.1:8000
paths:
  /api/v1/models/detect-protocol:
    post:
      description: Classify the endpoint's wire so the dialog needs no manual type selector.
      operationId: detect_protocol_api_v1_models_detect_protocol_post
      parameters:
      - in: header
        name: x-coffer-token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Coffer-Token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DetectProtocolIn'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetectProtocolOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: The request failed validation (`CONFIG_INVALID`); the submitted values are not echoed back.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Any other error, as Coffer's error envelope.
      summary: Detect Protocol
      tags:
      - models
  /api/v1/models/list-models:
    post:
      description: |-
        List the models a provider exposes (empty + message → enter manually).

        Each id comes back with a GUESSED modality so the connection's model table
        pre-fills a sensible kind; the user corrects it and the curated set stores
        the answer (spec provider-switching "Store a modality with each curated
        model").
      operationId: list_provider_models_api_v1_models_list_models_post
      parameters:
      - in: header
        name: x-coffer-token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Coffer-Token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListModelsIn'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderModelsOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: The request failed validation (`CONFIG_INVALID`); the submitted values are not echoed back.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Any other error, as Coffer's error envelope.
      summary: List Provider Models
      tags:
      - models
  /api/v1/models/test-connection:
    post:
      description: Probe a chat provider with a minimal request; 200 with ok=true/false.
      operationId: test_connection_api_v1_models_test_connection_post
      parameters:
      - in: header
        name: x-coffer-token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Coffer-Token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestConnectionIn'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestResultOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: The request failed validation (`CONFIG_INVALID`); the submitted values are not echoed back.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Any other error, as Coffer's error envelope.
      summary: Test Connection
      tags:
      - models
...
components:
  schemas:
...
    CuratedPrice:
      additionalProperties: false
      description: |-
        What the user says this connection charges for a model, in USD per
        million tokens (web search per thousand requests). Relays and resellers
        price differently from the vendor, so a connection's own price wins over
        the bundled snapshot when usage is costed (spec provider-switching "Price
        usage from a bundled snapshot and per-connection prices"). A cache category
        left out is charged at the input rate, so an estimate errs high.
      properties:
        cache_read:
          anyOf:
          - minimum: 0.0
            type: number
          - type: 'null'
          title: Cache Read
        cache_write_1h:
          anyOf:
          - minimum: 0.0
            type: number
          - type: 'null'
          title: Cache Write 1H
        cache_write_5m:
          anyOf:
          - minimum: 0.0
            type: number
          - type: 'null'
          title: Cache Write 5M
        input:
          minimum: 0.0
          title: Input
          type: number
        output:
          minimum: 0.0
          title: Output
          type: number
        web_search:
          anyOf:
          - minimum: 0.0
            type: number
          - type: 'null'
          title: Web Search
      required:
      - input
      - output
      title: CuratedPrice
      type: object
...
    DetectProtocolIn:
      properties:
        base_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Base Url
        secret_ref:
          anyOf:
          - type: string
          - type: 'null'
          title: Secret Ref
        secret_value:
          anyOf:
          - type: string
          - type: 'null'
          title: Secret Value
      title: DetectProtocolIn
      type: object
    DetectProtocolOut:
      properties:
        protocol:
          title: Protocol
          type: string
      required:
      - protocol
      title: DetectProtocolOut
      type: object
    ErrorDetail:
      properties:
        code:
          examples:
          - RESOURCE_NOT_FOUND
          title: Code
          type: string
        details:
          additionalProperties: true
          title: Details
          type: object
        message:
          examples:
          - 'resource not found: mcp_server:filesystem'
          title: Message
          type: string
      required:
      - code
      - message
      - details
      title: ErrorDetail
      type: object
    ErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
      required:
      - error
      title: ErrorResponse
      type: object
...
    ListModelsIn:
      properties:
        base_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Base Url
        provider:
          title: Provider
          type: string
        secret_ref:
          anyOf:
          - type: string
          - type: 'null'
          title: Secret Ref
        secret_value:
          anyOf:
          - type: string
          - type: 'null'
          title: Secret Value
      required:
      - provider
      title: ListModelsIn
      type: object
...
    Modality:
      description: The kind of output a model produces.
      enum:
      - text
      - embedding
      - image
      - video
      - audio
      title: Modality
      type: string
...
    ProviderModel:
      description: |-
        One model on a connection: an opaque id plus what KIND of model it is.

        The same shape is used both ways — the curated entries a connection stores
        and the ids endpoint introspection discovers — so the connection editor can
        round-trip a discovered model into the curated set without reshaping it.
        ``modality`` defaults to ``text``, the kind every curated set held before
        modalities existed.
      properties:
        context_window:
          anyOf:
          - maximum: 100000000.0
            minimum: 1024.0
            type: integer
          - type: 'null'
          title: Context Window
        default_effort:
          anyOf:
          - type: string
          - type: 'null'
          title: Default Effort
        effort_levels:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Effort Levels
        id:
          maxLength: 200
          minLength: 1
          title: Id
          type: string
        modality:
          $ref: '#/components/schemas/Modality'
          default: text
        price:
          anyOf:
          - $ref: '#/components/schemas/CuratedPrice'
          - type: 'null'
      required:
      - id
      title: ProviderModel
      type: object
    ProviderModelsOut:
      description: |-
        What an endpoint reports it serves, each id tagged with the modality
        Coffer INFERRED from its name — a pre-fill for the connection editor's
        model table, which the user corrects. Nothing downstream reads this guess:
        once an entry is curated, the stored modality is the truth.
      properties:
        message:
          default: ''
          title: Message
          type: string
        models:
          items:
            $ref: '#/components/schemas/ProviderModel'
          title: Models
          type: array
        reachable:
          default: true
          title: Reachable
          type: boolean
      required:
      - models
      - message
      - reachable
      title: ProviderModelsOut
      type: object
...
    TestConnectionIn:
      properties:
        base_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Base Url
        model:
          title: Model
          type: string
        provider:
          title: Provider
          type: string
        secret_ref:
          anyOf:
          - type: string
          - type: 'null'
          title: Secret Ref
        secret_value:
          anyOf:
          - type: string
          - type: 'null'
          title: Secret Value
      required:
      - provider
      - model
      title: TestConnectionIn
      type: object
    TestResultOut:
      properties:
        detail:
          additionalProperties: true
          default: {}
          title: Detail
          type: object
        message:
          title: Message
          type: string
        ok:
          title: Ok
          type: boolean
      required:
      - ok
      - message
      - detail
      title: TestResultOut
      type: object
...